+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 ## 0.12.14
 
 * Penalty for too large readme, changelog or example.
+
 * Penalty for too large `pubspec.yaml`.
+
+* Display calculated health and maintenance scores when `pana` is called with `--scores`.
 
 ## 0.12.13+1